refactor(rust): move old commands to a submodule
Refactor ockam_command crate to move depricated commands to a
submodule without changing their interface. We'll later remove them
when newer alternative are in place.
