+++ implementations/rust/ockam/ockam_command/src/old/session/error.rs
